requirements.txt
python-telegram-bot==13.13
python-dotenv==0.20.0
SQLAlchemy==1.4.39
psycopg2==2.9.3